hashing algo

+++ C++/algo/rolling-hash/main.cpp
@@ -18,9 +18,9 @@ int hashIt(string text){
     double NUM_OF_CHAR = 26;
     int PRIME_NUM = 5381;
     reverse(text.begin(),text.end());
-    float hash =0;
+    int hash =0;
     for(int i=0;i<text.length();i++){
-        double posNum = text[i] - 'A' + 1;
+        int posNum = text[i] - 'A' + 1;
         int posHash = posNum * pow(NUM_OF_CHAR,i);
         hash = fmod(hash + posHash,PRIME_NUM);
     }
